Bus: Add bus with round robin arbiter
This bus can have an arbitrary number of masters and up to ten slaves,
which are defined with a parameter. The signals are flat to allow the
variable number of participants. The memory map is also configured
with parameters. The round robin arbitration is an extra module.

For more details on usage see the verilog file.